requirements.txt
Flask==2.0.1                 # Framework web léger pour Python
mysql-connector-python==8.0.25  # Connecteur pour se connecter à MySQL depuis Python
werkzeug==2.0.3